Update system configuration and dependencies, and add new mail class
This commit includes several updates. The main changes are updating the system configuration files and upgrading the project's dependencies to the latest versions. In addition to this, the previous exclusion folders were removed and new ones were added to the IDE. We also added a new class for sending mail when an athlete is registered.