Add registration notification for athletes
In this update, a notification system for newly registered athletes is implemented. The AthleteRegistered notification class has been created and is called whenever a new athlete is created in the Athlete model. Also, minor UI changes have been made to provide a loading message in the registration form and the seeding of example athletes has been commented out.