Update route definitions to named routes
Changes were made to the route definitions throughout the web application. Instead of using URLs directly, Laravel's named routes were adopted. This was done for the Footer and Menu components, as well as updated in the 'routes/web.php' file. This approach results in more maintainable links and better organized code.