Refactor Athlete and Donator detail components
This change refactors AthleteDetails and DonatorDetails to use arrays instead of objects for data. This approach allows the explicit control over which data is passed to the UI. Instead of passing the entire object, only selected fields are now passed to improve data privacy and reduce redundancy. The attached view files were also updated to handle array data instead of objects.