Enable JavascriptInputFilledCheck and update form spam handling
This commit enables the JavascriptInputFilledCheck in the honey configuration. In addition, updates are made to the save methods in both the BecomeDonatorForm.php and BecomeAthleteForm.php files to change the way spam handling is implemented. Now, instead of aborting the save operation when spam is detected, an error message is added to alert about the detected spam.